[aml] changed internal time_point to time_point_ms so we can seek to milliseconds instead of seconds, changed message params to int64_t/double so we always have a 64 bit passing params. This really needs more work and all the other time vars (current_time, full_time, etc) also need to be changed to a millisecond timebase

+++ package/amlogic/libplayer/src/amplayer/player/include/message.h
@@ -69,16 +69,16 @@ typedef struct {
     union {
         char *filename;
         char *file_list;
-        int param;
-        float f_param;
+        int64_t param;
+        double f_param;
     };
     union {
-        int param1;
-        float f_param1;
+        int64_t param1;
+        double f_param1;
     };
     union {
-        int param2;
-        float f_param2;
+        int64_t param2;
+        double f_param2;
     };
 } player_cmd_t;
